Move Old CFG to .Sonar directory and namespace
